+++ lib/hall-native/hall.h
@@ -12,6 +12,6 @@ namespace hall {
         void init (uint8_t ms =100);
         void deinit ();
         auto millis () -> uint32_t;
-        auto micros () -> uint16_t;
+        auto micros () -> uint32_t;
     }
 }
